refactoring: move redundant functions in utils_streamlit_app
* Functions that moved are :
 * filtering_path_file()
 * load_numpy()
 * map_path_experiments_to_params()
 * selection_of_experiment()

* Documentation

the purpose is to make the code more modular and easy to read.
